+++ eoreader/products/optical/s2_e84_product.py
@@ -350,7 +350,7 @@ class S2E84Product(OpticalProduct):
             band_path,
             pixel_size=pixel_size,
             size=size,
-            resampling=Resampling.bilinear,
+            resampling=kwargs.pop("resampling", self.band_resampling),
             **kwargs,
         )