+++ eoreader/products/optical/s2_theia_product.py
@@ -373,7 +373,7 @@ class S2TheiaProduct(OpticalProduct):
             band_path,
             pixel_size=pixel_size,
             size=size,
-            resampling=Resampling.bilinear,
+            resampling=kwargs.pop("resampling", self.band_resampling),
             **kwargs,
         )