+++ eoreader/products/optical/s3_product.py
@@ -467,7 +467,7 @@ class S3Product(OpticalProduct):
             band_path,
             pixel_size=pixel_size,
             size=size,
-            resampling=Resampling.bilinear,
+            resampling=kwargs.pop("resampling", self.band_resampling),
             **kwargs,
         )