+++ eoreader/products/sar/sar_product.py
@@ -641,7 +641,7 @@ class SarProduct(Product):
             band_path,
             pixel_size=pixel_size,
             size=size,
-            resampling=Resampling.bilinear,
+            resampling=kwargs.pop("resampling", self.band_resampling),
             as_type=np.float32,
             **kwargs,
         )